Add debug mode and fix single player throw error on chest events

+++ Source/DataStructures/LootrChest.cs
@@ -90,6 +90,7 @@ public class LootrChest
 		if (!playerItems[player].All(item => item.IsAir))
 			return;
 
-		playerRestoreTime[player] = Main.GameUpdateCount + (LootrConfig.Instance.SecondsToRestore * 60);
+		//FIXME: change Main.GameUpdateTime on sth better!!!
+		playerRestoreTime[player] = Main.GameUpdateCount + (uint)(LootrConfig.Instance.SecondsToRestore * 60);
 	}
 }
